Add bootstrap playbooks.
When creating new servers the SSH fingerprint is not present in the
local ~/.ssh/known_hosts file, making it impossible to run any playbooks
in there. The bootstrap-local.yml file will gather the servers' SSH
fingerprint and add it to the local known_hosts file, enabling the
running of playbooks against them afterwards.

The Ubuntu images that we're using in DigitalOcean do not come with
python pre-installed, which makes it impossible to run any kind of
Ansible modules there. We have introduced the bootstrap-remote.yml
playbook which runs a raw command for installing python in those
servers.